requirements.txt
django==3.2.*
Pillow==8.4.*
django-admin-sortable2==1.0
django-tinymce==3.5.*
environs[django]==9.5.*
requests==2.28.*
